update package name for scaling-postgresql module.
